Make text replacement APIs respect the document's encoding
Notepad++ is *not* a .NET app: we can't assume every buffer is
UTF-8 and blindly convert every text stream to Unicode.

Notepad++ *is* a Unicode app, in the same way that Windows
is a Unicode OS. So we can still take Unicode for granted with
APIs that send text to GUI elements: autocompletion popups,
the line number margin, etc.